Refactor Onboarding component to use feature flags and hooks to its separate component.
