+++ app/client/src/layoutSystems/common/dropTarget/DropTargetComponent.tsx
@@ -23,7 +23,6 @@ import { useDispatch } from "react-redux";
 import { getDragDetails } from "sagas/selectors";
 import {
   combinedPreviewModeSelector,
-  getIsMobileCanvasLayout,
   getOccupiedSpacesSelectorForContainer,
 } from "selectors/editorSelectors";
 import { getCanvasSnapRows } from "utils/WidgetPropsUtils";
@@ -32,29 +31,16 @@ import { useShowPropertyPane } from "utils/hooks/dragResizeHooks";
 import { useWidgetSelection } from "utils/hooks/useWidgetSelection";
 import { calculateDropTargetRows } from "./DropTargetUtils";
 
-import { FEATURE_FLAG } from "@appsmith/entities/FeatureFlag";
-import {
-  EditorEntityTab,
-  EditorState as IDEAppState,
-} from "@appsmith/entities/IDE/constants";
-import { isAirgapped } from "@appsmith/utils/airgapHelpers";
 import { LayoutSystemTypes } from "layoutSystems/types";
-import {
-  useCurrentAppState,
-  useCurrentEditorState,
-} from "pages/Editor/IDE/hooks";
 import { getIsAppSettingsPaneWithNavigationTabOpen } from "selectors/appSettingsPaneSelectors";
 import { getLayoutSystemType } from "selectors/layoutSystemSelectors";
-import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import { getWidgetSelectionBlock } from "selectors/ui";
 import {
   isAutoHeightEnabledForWidget,
   isAutoHeightEnabledForWidgetWithLimits,
 } from "widgets/WidgetUtils";
 import DragLayerComponent from "./DragLayerComponent";
-import StarterBuildingBlocks from "./starterBuildingBlocks";
-import BuildingBlockExplorerDropTarget from "./buildingBlockExplorerDropTarget";
-import { isDraggingBuildingBlockToCanvas } from "selectors/buildingBlocksSelectors";
+import Onboarding from "./OnBoarding";
 
 export type DropTargetComponentProps = PropsWithChildren<{
   snapColumnSpace: number;
@@ -77,61 +63,6 @@ const StyledDropTarget = styled.div`
   user-select: none;
   z-index: 1;
 `;
-
-function Onboarding() {
-  const isMobileCanvas = useSelector(getIsMobileCanvasLayout);
-  const isDraggingBuildingBlock = useSelector(isDraggingBuildingBlockToCanvas);
-  const appState = useCurrentAppState();
-  const isAirgappedInstance = isAirgapped();
-  const { segment } = useCurrentEditorState();
-
-  const showStarterTemplatesInsteadOfBlankCanvas = useFeatureFlag(
-    FEATURE_FLAG.ab_show_templates_instead_of_blank_canvas_enabled,
-  );
-  const releaseDragDropBuildingBlocksEnabled = useFeatureFlag(
-    FEATURE_FLAG.release_drag_drop_building_blocks_enabled,
-  );
-
-  const shouldShowStarterTemplates = useMemo(
-    () =>
-      showStarterTemplatesInsteadOfBlankCanvas &&
-      !isMobileCanvas &&
-      !isAirgappedInstance &&
-      !releaseDragDropBuildingBlocksEnabled, // Hide starter templates when drag-drop building blocks are available
-    [
-      showStarterTemplatesInsteadOfBlankCanvas,
-      isMobileCanvas,
-      isAirgappedInstance,
-      releaseDragDropBuildingBlocksEnabled,
-    ],
-  );
-
-  const shouldShowBuildingBlocksDropTarget = useMemo(
-    () => releaseDragDropBuildingBlocksEnabled && !isDraggingBuildingBlock,
-    [releaseDragDropBuildingBlocksEnabled, isDraggingBuildingBlock],
-  );
-
-  const isEditorState = appState === IDEAppState.EDITOR;
-  const isUISegment = segment === EditorEntityTab.UI;
-
-  if (shouldShowStarterTemplates && isEditorState) {
-    return <StarterBuildingBlocks />;
-  } else if (
-    shouldShowBuildingBlocksDropTarget &&
-    isEditorState &&
-    isUISegment
-  ) {
-    return <BuildingBlockExplorerDropTarget />;
-  } else if (!isDraggingBuildingBlock) {
-    return (
-      <h2 className="absolute top-0 left-0 right-0 flex items-end h-108 justify-center text-2xl font-bold text-gray-300">
-        Drag and drop a widget here
-      </h2>
-    );
-  } else {
-    return null;
-  }
-}
 
 /*
   This context will provide the function which will help the draglayer and resizablecomponents trigger
